+++ src/cxxstd.rs
@@ -111,8 +111,7 @@ impl<'a> CxxStringRef<'a> {
 
     /// 長さを取得する。
     pub fn len(&self) -> usize {
-        let len = unsafe { ffi::std_string_size(self.as_ptr()) };
-        len.max(0) as usize
+        unsafe { ffi::std_string_size(self.as_ptr()) }
     }
 
     /// 空かどうかを返す。